Create NoseCone_HaackSeries.py
Description - Takes in the arguments like the height of the nose cone and the Haack constant for the Hack series, then 
creates a spline as a sketch that fits the curve geometry. Then an offset curve is added to the sketch 
Which is then closed to make an area enclosing a curve. The enclosed area is then finally revolved around the axis.